requirements.txt
torch>=2.2
torchvision>=0.17
timm>=0.9.12
numpy>=1.26
pillow>=10.2
opencv-python>=4.9
matplotlib>=3.8
flask>=3.0
flask_sqlalchemy>=3.1
python-dotenv>=1.0
tqdm>=4.66
scikit-learn>=1.4
pydantic>=2.7
# Optional LLM dependency (comment out if not using):
openai>=1.30
google-generativeai